near complete rewrite
This scraper checks the first post title's timestamp and then checks each proceeding post title's timestamp looking for a timestamp that doesn't contain 'days'. When it finds that we know we've located the hottest post that isn't pinned or over 24 hours old.
We scrape that data as well as any upvotes (converted to numbers) into a JSON file. If the upvotes are over 500 we have Hot Post text appended to the json dictionary