Imported Boto3,  changed exception handling, connections to json files in bucket, added hidden settings
Starting at the top with this one we at boto3 and AWS hidden settings from settings file, allowing us to connect to our json files in the AWS bucket S3 storage. lines 15-30

Next we add in the  code to write to the json files in bucket, as well as minor text changes indicating patent change. lines 69/70

Fixed a link issue not specifying to click the anchor of the link "/a" line 77
Finally we had our exceptions not printing out the exception which was silly. Fixed that, also Log the exception now.