changed secretKeys location
(in Linux) Was unable to import secret keys from settings.py in the Tesla file/module  in a sibling folder. Ex TeslaSpectra/Selenium vs TeslaSpectra/Tesla +__init__.py.
But was able to import the file with  __init__.py in a child folder. Ex TeslaSpectra/Selenium/folder with .py and __init__.py file.