added href link to scrape, + as a dictionary variable, update 'Hot' Status
Scrape the associated link to the post and paste it as part of the website. If there are more than 4000 votes (used to be 1000) then its a "Hot" post. also if its not hot its '' which returns False in the Boolean based If statement  now used in templates/index.html