Xpath changes and added exit button click section
a dialog box was showing up asking for confirmation if user is in canada, making the script not work. The button to exit out of this box is dealt with on lines 38 and 39 through element.click().

I also simplified the XPaths listed for model S price and model S Plaid prices lines 41-45