TheLimitingFactor's Youtube channel is scraped for new content
Basically, go to the webpage of the channel, scrape the title and the link to the latest video, scrape the location to click the link in selenium, as well as the views and the post date of the video. click the video link, then the share button, then scrape the embedButton location. Click that and then scrape the embedcode. Put that altogether in a dictionary, open some json files to store the content. One in S3 one locally