moved Try: clause, limited range of runtime, fixed broken xpath to postlink
The Try: clause should start at the beginning of the  function so that the catch exceptions line is run, ensuring driver.quit() occurs.

The while True: loop has the potential to run continuously if "hours" is not found, so we need to make sure there is a limit to how many posts are checked for the word "hours", which is done with for postNumber in range(0, 10);

The last part to fix was a broken xpath for postLink.