remove unused code that goes to currently broken links, remove old bootstrap ref.
The {if user.is_authenticated} function was created to see if I could apply logic to what the available login buttons would be if someone wanted  to login and were a member vs a first time visitor, but went to broken links. I removed this code for now to avoid the broken link issue.

Changed navbar to .nav css class

Also removed column width type for title