Update index.html
this index.html has been updated to display tesla's most recent patent filing(S) depending on whether more than 1 (up to 5) was scraped last. Because the script which populates the json files only scrapes up to 4 patents, line 31 should probably say {% for i in "1234" %}. ChatGPT was a big help in formatting this code, since it was difficult to find the correct syntax for this for loop.

on line 33 The if loop checks of the patent key is present before creating the next jumbotron, ensuring only as many jumbotrons are created as there are "patent" entries in the dictionary.

I commented out the limiting factor section because the resizing of the jumbotron with the video embed seems to be messing with the header. working on a fix