replaces the Selenium based reddit scraper
this script uses the RedditApi to retrieve subreddit details for the hot posts. It is much simpler and faster than the previous scraper script which relied upon selenium. about 4X faster. I also simplified the "hotPost" if else statements used to determine if a post was particularly "hot" or had a large number of upvotes.