created a new dhango template which stores scraped subreddit details
the meta tags from ChatGPT, which also helped with the for loop syntax.
this page loops through the dictionary values for each channel and creates a jumbotron with the details of the post  and a link to it. There is also a (now working) if loop which will show a little fire image if the post is "hot"