Various updates
Line 9- Added reference to js.cookie.js for use of the cooke.js library in auto-identifying site cookies.
Line 11- Added if statement using accepted_cookies as a boolean value to determine if we load the google adsense cookies.
Line 41- Added a cookie banner with id's associated to Jquery functions
Line 60- Added cookiebanner.js for use of the cookie banner