Changed new_scrape to already_scraped
Instead of a boolean new_scrape value
we replaced this value with an integer already_scraped which will be used to check if there was a newly scraped video for each of the channels. for example if already_scraped =3 and channels =3 no new video has been scraped, therefore end program.

Otherwise if the video details aren't empty then upload those details to the storage in S3.