Made the script more robust with try clauses, wait functions
Added in try, wait, with exception clauses for certain parts of the scraper to make it more robust. Some patents dont have abstracts for example, and sometimes the page loading time varies past 30 seconds, these updates ensure 100 seconds of page load time both for when we click the search button and when we go to scrape the patent document provided later in the script. Also added exceptions for error handling.